% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sf_GO.R
\name{Sf_GO}
\alias{Sf_GO}
\title{GO enrichment of Spodoptera frugiperda}
\usage{
Sf_GO(genelist, method = p.adjust.methods, withanc = TRUE)
}
\arguments{
\item{genelist}{should be factors}

\item{method}{will be set to "holm" as default if not chosen}

\item{withanc}{means that genes will be enriched into ancestor GO terms}
}
\value{
GO enrichment Table
}
\description{
To enrich genes into Sf genome (from BIPAA).
}
\author{
Pai Li
}
