% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tair_GO.R
\name{Tair_GO}
\alias{Tair_GO}
\title{GO enrichment of Arabidopsis thaliana}
\usage{
Tair_GO(genelist, method = p.adjust.methods, withanc = TRUE)
}
\arguments{
\item{genelist}{should be factors}

\item{method}{will be set to "holm" as default if not chosen}

\item{withanc}{means that genes will be enriched into ancestor GO terms}
}
\value{
GO enrichment Table
}
\description{
To enrich genes into TAIR genome.  It'll be updated every single month.
}
\author{
Pai Li
}
